Move test that stops the batcher to its own package.
