+++ op-acceptance-tests/tests/interop/proofs/serial/interop_fault_proofs_test.go
@@ -1,4 +1,4 @@
-package proofs
+package serial
 
 import (
 	"testing"
